Makes all Places singleton and all Presenters non-singleton in sample
Places are a lightweight reference to the presenter, so a cached forever is reasonable. But presenter are heavyweight instances, even more the view that it has, so it must be discarded. Currently it is discarded and created too often, some timeout caching mechanism should be implemented or be able to apply by library users.